% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  databaseName = "Unknown",
  createCohorts = TRUE,
  runPrescriptionNum = TRUE,
  runDUR = TRUE,
  resultsToZip = TRUE,
  yearStartDate = as.Date("2006-01-01"),
  yearEndDate = as.Date("2022-12-31"),
  monthStartDate = as.Date("2006-01-01"),
  monthEndDate = as.Date("2022-12-31")
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseName}{The name of the database (e.g. 'SynPUFs').}

\item{createCohorts}{Create the cohortTable table with the exposure and outcome cohorts?}

\item{runPrescriptionNum}{Calculate the monthly prescription numbers?}

\item{runDUR}{Calculate the monthly prescription numbers of DUR?}

\item{resultsToZip}{Create the zip files?}

\item{yearStartDate}{Start date of observation period, has to be on or later than 2006-01-01}

\item{yearEndDate}{End date of observation period, has to be on or earlier than the latest date in the database}

\item{monthStartDate}{Start date for monthly prescription numbers, has to be on or later than 2006-01-01}

\item{monthEndDate}{End date for monthly prescription numbers, has to on or earlier than the latest date in the database}
}
\description{
Execute the Study
}
\details{
This function executes the SAGE Study.

The \code{createCohorts}, \code{createAllCohorts}, \code{computePrescriptionNum} arguments
are intended to be used to run parts of the full study at a time, but none of the parts are considered to be optional.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "redshift",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        cohortDatabaseSchema = "study_results",
        cohortTable = "cohort",
        outputFolder = "c:/temp/study_results")
}

}
